Use u32 instead of u16 for extButton so constants don't collide

+++ include/dolphin/pad.h
@@ -101,7 +101,7 @@ typedef struct PADStatus {
   u8 analogB;
   s8 err;
 #ifdef TARGET_PC
-  u16 extButton;
+  u32 extButton;
 #endif
 } PADStatus;
